// src/app/core/api/model/clusterDnsZoneResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DnsZoneResponseDto } from './dnsZoneResponseDto';


export interface ClusterDnsZoneResponseDto { 
    id: string;
    clusterId: string;
    dnsZoneId: string;
    dnsZone: DnsZoneResponseDto;
    certificateProvider?: ClusterDnsZoneResponseDto.CertificateProviderEnum;
    acmeEmail?: string;
    wildcardCertificate: boolean;
    reconciliationStatus: ClusterDnsZoneResponseDto.ReconciliationStatusEnum;
    lastReconciliationAt?: string;
    errorMessage?: string;
    createdAt: string;
    updatedAt: string;
}
export namespace ClusterDnsZoneResponseDto {
    export const CertificateProviderEnum = {
        LetsEncrypt: 'lets_encrypt',
        LetsEncryptStaging: 'lets_encrypt_staging'
    } as const;
    export type CertificateProviderEnum = typeof CertificateProviderEnum[keyof typeof CertificateProviderEnum];
    export const ReconciliationStatusEnum = {
        Pending: 'PENDING',
        InSync: 'IN_SYNC',
        Drift: 'DRIFT',
        Reconciling: 'RECONCILING',
        Error: 'ERROR'
    } as const;
    export type ReconciliationStatusEnum = typeof ReconciliationStatusEnum[keyof typeof ReconciliationStatusEnum];
}