// src/app/core/api/model/assignDnsZoneDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AssignDnsZoneDto { 
    /**
     * ID of the DnsZoneEntity to assign to the cluster
     */
    dnsZoneId: string;
    /**
     * Certificate provider for TLS
     */
    certificateProvider?: AssignDnsZoneDto.CertificateProviderEnum;
    /**
     * Email for ACME certificate registration
     */
    acmeEmail?: string;
    /**
     * Whether to use a wildcard certificate
     */
    wildcardCertificate?: boolean;
}
export namespace AssignDnsZoneDto {
    export const CertificateProviderEnum = {
        LetsEncrypt: 'lets_encrypt',
        LetsEncryptStaging: 'lets_encrypt_staging'
    } as const;
    export type CertificateProviderEnum = typeof CertificateProviderEnum[keyof typeof CertificateProviderEnum];
}